refactor: move test for account balance to the base tests for account entity
